requirements.txt
fastapi
uvicorn[standard]
python-dotenv
faster-whisper
